fqdn: move the NameManager into its own package
This is a preparation change to modularising the name manager into its
own cell. Pulling it out from the fqdn package will allow depending on
the endpointmanager package, which would have otherwise caused an import
cycle. The cycle would have been fqdn -> epmgr -> ep -> fqdn, but with
pulling the name manager out of fqdn proper, we can have it depend both
on fqdn and the ep packages.
