--- pkg/fqdn/helpers_test.go
@@ -1,85 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-// Copyright Authors of Cilium
-
-package fqdn
-
-import (
-	"net/netip"
-	"testing"
-	"time"
-
-	"github.com/sirupsen/logrus"
-	"github.com/stretchr/testify/require"
-
-	"github.com/cilium/cilium/pkg/ip"
-	"github.com/cilium/cilium/pkg/policy/api"
-)
-
-var (
-	ciliumIOSel = api.FQDNSelector{
-		MatchName: "cilium.io",
-	}
-
-	githubSel = api.FQDNSelector{
-		MatchName: "github.com",
-	}
-
-	ciliumIOSelMatchPattern = api.FQDNSelector{
-		MatchPattern: "*cilium.io.",
-	}
-)
-
-func TestMapIPsToSelectors(t *testing.T) {
-	var (
-		ciliumIP1   = netip.MustParseAddr("1.2.3.4")
-		ciliumIP2   = netip.MustParseAddr("1.2.3.5")
-		nameManager = NewNameManager(Config{
-			MinTTL: 1,
-			Cache:  NewDNSCache(0),
-		})
-	)
-
-	log.Level = logrus.DebugLevel
-
-	// Create DNS cache
-	now := time.Now()
-	cache := nameManager.cache
-
-	// Empty cache.
-	nameIPMapping := nameManager.mapSelectorsToNamesLocked(ciliumIOSel)
-	require.Empty(t, nameIPMapping)
-
-	// Just one IP.
-	ciliumIOName := prepareMatchName(ciliumIOSel.MatchName)
-	changed := cache.Update(now, ciliumIOName, []netip.Addr{ciliumIP1}, 100)
-	require.True(t, changed)
-	nameIPMapping = nameManager.mapSelectorsToNamesLocked(ciliumIOSel)
-	require.Len(t, nameIPMapping, 1)
-	println(ciliumIOSel.MatchName)
-	ciliumIPs, ok := nameIPMapping[ciliumIOName]
-	require.True(t, ok)
-	require.Len(t, ciliumIPs, 1)
-	require.Equal(t, ciliumIP1, ciliumIPs[0])
-
-	// Two IPs now.
-	changed = cache.Update(now, ciliumIOName, []netip.Addr{ciliumIP1, ciliumIP2}, 100)
-	require.True(t, changed)
-	nameIPMapping = nameManager.mapSelectorsToNamesLocked(ciliumIOSel)
-	require.Len(t, nameIPMapping, 1)
-	ciliumIPs, ok = nameIPMapping[ciliumIOName]
-	require.True(t, ok)
-	require.Len(t, ciliumIPs, 2)
-	ip.SortAddrList(ciliumIPs)
-	require.Equal(t, ciliumIP1, ciliumIPs[0])
-	require.Equal(t, ciliumIP2, ciliumIPs[1])
-
-	// Test with a MatchPattern.
-	nameIPMapping = nameManager.mapSelectorsToNamesLocked(ciliumIOSelMatchPattern)
-	require.Len(t, nameIPMapping, 1)
-	ciliumIPs, ok = nameIPMapping[ciliumIOName]
-	require.True(t, ok)
-	require.Len(t, ciliumIPs, 2)
-	ip.SortAddrList(ciliumIPs)
-	require.Equal(t, ciliumIP1, ciliumIPs[0])
-	require.Equal(t, ciliumIP2, ciliumIPs[1])
-}
